wip: testbench, locations, display hack, either
display hack: when guard errors occur, log the error message with
`Display` if an impl exists, falling back to `Debug` otherwise

item locations: log the file and line for catchers and routes

either: impls `FromParam` for `Either`. changes the implementation of
`FromParam` for all `T: FromStr` to return the actual error when the
conversion fails instead of the previous `&str` for the value that
failed to parse. This results in better log error messages on guard
failure. To regain the ability to recover the value that failed to
parse, `FromParam` was implemented for `Either<A, B>`; `Either<T, &str>`
can be used as the previous `Result<T, &str>` could, with the benefit
that it now works for all `T`. `FromUriParam` and `UriDisplay` were also
implemented for `Either<A, B>`.

testbench: fix the testbench for new logging and test that display hack
works as intended for both `T: Display` and `T: !Display + Debug`.